% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interrogators.R
\name{binomial_evaluator}
\alias{binomial_evaluator}
\title{Evaluator for binomial tests}
\usage{
binomial_evaluator(trials1, estimates1, trials2, estimates2)
}
\arguments{
\item{trials1}{vector with number of trials for sample 1 (with bookings)}

\item{estimates1}{vector with estimated binomial probabilities for sample 1 (with bookings)}

\item{trials2}{vector with number of trials for sample 2 (without bookings)}

\item{estimates2}{vector with estimated binomial probabilities for sample 2 (without bookings)}
}
\value{
fraction of correct guesses
}
\description{
Evaluator for binomial tests
}
\examples{
trials1 = c(10,10)
estimates1 = c(0.5, 0.5)
trials2 = c(10,10)
estimates2 = c(0.4, 0.6)
binomial_evaluatortrials1, estimates1, trials2, estimates2)
}
