% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_generators.R
\name{binomial_systematic_generator}
\alias{binomial_systematic_generator}
\title{Generate samples of pairs to be tested}
\usage{
binomial_systematic_generator(d, trial_column, success_column, estimate_columns,
  min_observations = 5, combine_low_volume_items = T,
  min_sample_size = 1000)
}
\arguments{
\item{d}{data frame with conversion data}

\item{trial_column}{column name with number of trials}

\item{success_column}{column name with number of successes}

\item{estimate_columns}{column names with conversion estimates}

\item{min_observations}{minimum number of observations per item}

\item{combine_low_volume_items}{flag indicating whether items that have less observations than minimum threshold should be combined}

\item{min_sample_size}{minimum number of random samples (and tests) generated}
}
\value{
list with two data frames, a sample of converted elements, and a sample (of the same size) with non converted elements
}
\description{
Generate samples of pairs to be tested
}
