% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combiners.R
\name{combine_all_binomial}
\alias{combine_all_binomial}
\title{Combine low observation items in binomial dataset}
\usage{
combine_all_binomial(d, trial_column, success_column, estimate_columns,
  use_groups = FALSE)
}
\arguments{
\item{d}{data.frame}

\item{trial_column}{column name with number of trials}

\item{success_column}{column name with number of successes}

\item{estimate_columns}{column names with conversion estimates}

\item{use_groups}{use 'groups' column to combine itemsg}
}
\value{
combined data.frame
}
\description{
Combine low observation items in binomial dataset
}
