% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluators.R
\name{continuous_evaluator}
\alias{continuous_evaluator}
\title{Evaluator for continuous quantities}
\usage{
continuous_evaluator(observed1, estimates1, observed2, estimates2)
}
\arguments{
\item{observed1}{vector with values of observed quantities for sample 1}

\item{estimates1}{vector with values of estimated quantities for sample 1}

\item{observed2}{vector with values of observed quantities for sample 2}

\item{estimates2}{vector with values of estimated quantities for sample 2}
}
\value{
percent of correct guesses
}
\description{
Evaluator for continuous quantities
}
\examples{
obs1 <- c(100, 180)
obs2 <- c(150, 150)
est1 <- c(90, 160)
est2 <- c(120, 170)
continuous_evaluator(obs1, est1, obs2, est2)
}
