% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{continuous_turing_error}
\alias{continuous_turing_error}
\title{Turing prediction error for continuous data}
\usage{
continuous_turing_error(d, weight_column, obs_column, estimate_columns,
  min_weight = NULL, sample_size = 1000)
}
\arguments{
\item{d}{data frame with data}

\item{estimate_columns}{names of the columns that have conversion estimates}

\item{trial_column}{name of the column with number of trials}

\item{success_column}{name of column with number of successes}

\item{min_observations}{minimum number of observations per item}

\item{combine_low_volume_items}{combine items that have less observations than minimum threshold}

\item{min_sample_size}{minimum number of random samples (and tests) generated}
}
\value{
turing error
}
\description{
Turing prediction error for continuous data
}
