% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{conversion_turing_error}
\alias{conversion_turing_error}
\title{Probabilistic Turing Error for conversion data}
\usage{
conversion_turing_error(d, trial_column, success_column, estimate_columns,
  min_observations = 5, combine_low_volume_items = T,
  min_sample_size = 1000)
}
\arguments{
\item{d}{data frame with conversion data}

\item{trial_column}{column name with number of trials}

\item{success_column}{column name with number of successes}

\item{estimate_columns}{column names with conversion estimates}

\item{min_observations}{minimum number of observations per item}

\item{combine_low_volume_items}{flag indicating whether items that have less observations than minimum threshold should be combined}

\item{min_sample_size}{minimum number of random samples (and tests) generated}
}
\value{
probabilistic turing error of conversion data
}
\description{
Probabilistic Turing Error for conversion data
}
