% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{continuous_turing_error}
\alias{continuous_turing_error}
\title{Probabilistic Turing Error for continuous data}
\usage{
continuous_turing_error(d, weight_column, obs_column, estimate_columns,
  min_weight = NULL, min_sample_size = 1000)
}
\arguments{
\item{d}{data frame with data}

\item{weight_column}{name of the column with weights per item}

\item{obs_column}{name of column with observed (continuous) values}

\item{estimate_columns}{names of the columns that have predicted values}

\item{min_weight}{minimum weight per item}

\item{min_sample_size}{minimum number of random samples (and tests) generated}
}
\value{
turing error
}
\description{
Probabilistic Turing Error for continuous data
}
